backend/pyproject.toml
[tool.poetry]
name = "clickup-backend"
version = "0.1.0"
description = "FastAPI backend for ClickUp clone"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.110.0"
uvicorn = "^0.27.1"
sqlmodel = "^0.0.14"
python-dotenv = "^1.0.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.1"
pydantic = {extras = ["email"], version = "^2.6.1"}
pydantic-settings = "^2.1.0"
email-validator = "^2.1.0.post1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
black = "^24.1.1"
isort = "^5.13.2"
mypy = "^1.8.0"
ruff = "^0.2.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

[tool.ruff]
line-length = 88
target-version = "py310"
select = ["E", "F", "B", "I"]
ignore = [] 